requirements.txt
# Base dependencies
gradio==4.12.0
librosa==0.10.1
numpy==1.26.4
torchvision==0.16.2
torchaudio==2.2.2
soundfile==0.12.1
psutil
librosa